@@ -6,28 +6,56 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "postinstall": "prisma generate"
   },
   "dependencies": {
-    "@clerk/nextjs": "^5.0.12",
+    "@ai-sdk/openai": "^0.0.14",
+    "@clerk/nextjs": "^5.1.2",
+    "@prisma/client": "^5.14.0",
+    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
+    "@tanstack/react-query": "^5.37.1",
+    "@tanstack/react-query-devtools": "^5.37.1",
+    "@uploadthing/react": "^6.5.4",
+    "@vercel/functions": "^1.0.2",
+    "ai": "^3.1.13",
+    "axios": "^1.7.2",
+    "babel-plugin-react-compiler": "^0.0.0-experimental-592953e-20240517",
+    "cheerio": "^1.0.0-rc.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "crawler-request": "^1.2.2",
+    "fs": "^0.0.1-security",
     "lucide": "^0.378.0",
     "lucide-react": "^0.378.0",
-    "next": "14.2.3",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^15.0.0-rc.0",
+    "next-themes": "^0.3.0",
+    "pdf-parse": "^1.1.1",
+    "pdf2json": "^3.1.2",
+    "react": "^19.0.0-rc-4c2e457c7c-20240522",
+    "react-dom": "^19.0.0-rc-4c2e457c7c-20240522",
+    "react-icons": "^5.2.1",
+    "react-pdftotext": "^1.2.0",
+    "react-reveal": "^1.2.2",
+    "sonner": "^1.4.41",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "tailwindcss-animated": "^1.1.0",
+    "uploadthing": "^6.12.0",
+    "vaul": "^0.9.1",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
-    "eslint-config-next": "14.2.3",
+    "eslint-config-next": "^15.0.0-rc.0",
     "postcss": "^8",
+    "prisma": "^5.14.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
