@@ -3,13 +3,17 @@ import { Manrope } from "next/font/google";
 import "./globals.css";
 import { cn } from "@/lib/utils";
 import Header from "@/components/header";
-import Footer from "@/components/footer";
+import Providers from "@/lib/providers";
+import { NextSSRPlugin } from "@uploadthing/react/next-ssr-plugin";
+import { extractRouterConfig } from "uploadthing/server";
+import { ourFileRouter } from "./api/uploadthing/core";
+import { Toaster } from "@/components/ui/sonner";
 
 const manrope = Manrope({ subsets: ["latin"], variable: "--font-sans" });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "myresume.guru",
+  description: "Land your dream job with myresume.guru",
 };
 
 export default function RootLayout({
@@ -18,17 +22,20 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
-      <body
-        className={cn(
-          "min-h-secreen bg-background font-sans antialiased",
-          manrope.variable
-        )}
-      >
-        <Header />
-        {children}
-        <Footer />
-      </body>
-    </html>
+    <Providers>
+      <html lang="en">
+        <body
+          className={cn(
+            " min-h-svh h-screen bg-white font-sans antialiased",
+            manrope.variable
+          )}
+        >
+          <NextSSRPlugin routerConfig={extractRouterConfig(ourFileRouter)} />
+          <Header />
+          <div className="flex w-full flex-col h-auto min-h-screen pt-20">{children}</div>
+          <Toaster />
+        </body>
+      </html>
+    </Providers>
   );
 }
